Add explicit meta information to home route
We were previously letting /'s meta data come from index.html. However, if you navigated to a page with alternate meta data and the navigated back to home, the meta data would not be restored to the original version. This fixes that problem at the cost of having to maintain the base meta information in two places (index.html and Home.js).