Improve screen janky screen transition animations
The CSSTransition component that powers our screen transitions does not handle React.Fragments properly. Instead of wrapping the input in some container, it applies its transition classes to only the top element.

As a workaround, this replaces our Fragments with divs so that CSSTransition has a container to attach classes to.

Additionally, this limits the height of the exiting scene to the height of the view port. If we don't do this, then there is a small window where the user can scroll past the content of the entering screen and see exiting content if the exiting scene is longer.