HHGDemo: Fix intermittent no playback in chrome
autoPlay doesn't always fire on Chrome because the "muted" attribute is not explicitly written to the dom.

On mount, we ensure muted is enabled, trigger playback, and then immediately pause the video. This should enable our onCanPlayThrough / animate in handlers to properly trigger playback when the video asset is both visible and loaded.